"Added NutrientModule and updated related functionality"
This commit includes the addition of the new NutrientModule into the DotNetServer application. The NutrientModuleExtension has been adjusted to a static class for creating its services and wired up in the main program. Necessary using statements were also added. In order to operate with the new module, NutrientRepository and UserRepository have seen changes in their contexts.

Additionally, some methods were updated and a new method was added in NorwegianFoodService to enhance cooperation with the new NutrientModule.

The NutrientRequest model was improved with methods to facilitate the information transfer from NorwegianFoodModule.

A NutrientsController was set up inside NutrientModule, and it provides methods to get, post, and delete nutrients by their id.

Commit also includes some minor changes like the cleaning code in csproj file, and creation of additional validations and controllers folders for NutrientModule.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>