Add MacroChart, refactor UserContentCharts, amend styling and button naming.
This commit introduces a new chart "MacroChart" and exports it in the index.ts. In "UserContentCharts.tsx" a significant refactor happened. The different macros are now elements of generic type "IGenericMacroDataChart". For each type, the specific data is extracted and given to the new "MacroChart" in return for a visually represented chart. In terms of styling, the NavBar is now sticky to the top and the unused button was removed. Also the "GenericCancelButton" in the "LoginForm.tsx" got renamed to "CancelBackPreviousRouteButton".

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>