Refactor NavBar layout and adjust link styles
Simplified code in NavBar.tsx by removing unnecessary 'button' tags around 'NavLink' component instances. This results in a cleaner structure and improved reusability of the code. Also, altered text color styles and hover behaviors on the NavLink components for better visual distinction and user experience. As part of alignment in routing change, the previously /user-food endpoint was renamed to /food for better accuracy and simplicity. This alteration was updated in App.tsx accordingly.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>