Change ChartPage to NewFeaturePage in application layout
In this commit, the 'ChartPage' webpage is replaced with a 'NewFeaturePage'. This change was done in the NavBar component as well as the indexed exports in pages folder. As reflected in the updated content of 'NewFeaturePage', the previous data-related functionalities have been removed.

Currently, 'NewFeaturePage'only displays an "Under Construction" message. This change aligns with our plan to incorporate a new feature into the application, hence the existing 'ChartPage' has been renamed and prepared for future updates.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>