Update application to fetch food list from server
This commit updates both the server and client applications so that the food list used on the client can be served by the server. The react search library Fuse.js is used to search for food in a list of 2000 foods from src/components/SearchBar.tsx. This list is then replaced by a new food list from the server obtained at its /foods endpoint in ../DotNetServer/FoodSeedMySqlDatabase/Program.cs, which is then used on React components via client/src/utils/ListOfDbFoods.ts.

The advantages of this update are fetched data are always up-to-date. Moreover, server resources are used more efficiently, and the client app is optimized because it no longer needs to load a large list of foods during startup. As a result, the UX for the user is enhanced.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>