Refactor display of chart dates and improve user actions interface
The changes reflect necessary refactorings for better UI/UX. The first change is the removal of a misleading comment on month retrieval from Date object in SingleMacroChart component, to prevent any future misunderstanding of the presented data. This change is based on the standard JavaScript Date handling where months are indeed 0-based, but the already existing +1 adjustment handles this, hence the comment was superfluous.

Furthermore, UI components were also revised for better coherence with the user actions context. This improvement aims to deliver a more intuitive and user-friendly interface, rendering actions, and data visualization clearer for the users.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>