Remove 'AboutPage' and refactor imports for project simplification
Removed the 'AboutPage' and 'FetchUserContent' files due to them being unused. Updated the 'FoodItem' import from the 'types' directory to the new 'Model' directory in multiple files as part of a project restructure. This change also applies to the imports of 'GetMacrosAndEnergy', 'IGenericMacroDataChart', 'UserIdResponse', and 'UserRequest'.

Moved the 'FetchUser' file under the 'Axios' directory to match project restructure. Also removed unused code and unnecessary variables which allows for more effective maintenance and improved cleanliness of the codebase.

Overall, these changes to the project structure are made to simplify the future maintainability and improve readability of the code.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>