Simplify navigation to SignUp in LoginForm component
Replaced the SignUp button with a Link in the LoginForm component. This change was made to simplify the navigation to the SignUp page. Previously, a state variable 'isSignUp' was being used to determine whether to navigate or not, that logic was removed and the button was replaced with a 'Link' to signup page for cleaner, more understandable code.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>