Implement user secrets for secure data storage
We've moved sensitive data (e.g. connection strings) from appsettings.json to user secrets for a more secure environment. We have added UserSecretsId in the project file and used the AddUserSecrets method in Program.cs for the environment configuration. We've opted to use user secrets as they are stored outside the project tree which enhances the security of the app, especially in production mode.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>