Update button styles and enhance user-interaction
Updated the color scheme of the 'Log Out' button to better match the app's overall design. Also, modified the placeholder text in the 'Meal' component from 'Search for a food...' to more generic 'Search' and 'Food' to serve as better cues for users. Search results rendered as clickable buttons instead of list items for better user-interaction. Enhanced the Landing page with a revised mission statement for better user understanding of the app's purpose.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>