Update UI and Backend code
Various UI fixes and enhancements, such as changes to button hover styles for better visibility and more intuitive interactions. Also included are overall color scheme adjustments to improve contrast and readability. The Parallax component was also enhanced with updated text info and better text alignment.

Backend modifications were also made, such as the removal of unnecessary components and code simplification. For example, the Unused context import was removed from UsersController and the unnecessary test route was removed for cleanup. Index was added to the Nutrient.cs to improve query performance.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>