Remove ProductionNorwegianFoodDbContext and update context setup
Removed the ProductionNorwegianFoodDbContext completely and updated the way contexts are added in the CoreModuleExtension to simplify the code and improve modularity. The NorwegianFoodDbContext is now used for both production and development environments, thereby maintaining uniformity. Additionally, data seed operations involving CsvHelper have been commented out in NorwegianFoodDbContext to prevent automatic data seeding on context instantiation. This is necessary for keeping the database setup flexible for different deployment circumstances.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>