Update URLs to point to Azure hosted API
Changed URLs used in various axios requests throughout the client side to point to the Azure hosted server API instead of the localhost server. This was done to prepare for production deployment and ensure all server calls will hit the appropriate endpoint. The CORS policy on the server side was also adjusted to allow requests from any origin to improve application accessibility.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>