Added new logo and updated the way it's used in Footer and NavBar
A new small bird logo was added to `tailwind.config.js` to optimize the website performance. The direct import of the image in `Footer.tsx` and `NavBar.tsx` components has been replaced with a reference to a TailwindCSS utility class that uses the logo from the configuration. This is to leverage CSS and optimize the website performance. A direct import of the image was removed from `assets/index.ts` because it's not used anymore.

Also, new files `vite.svg`, `index.html`, and `index-e768f981.css` were added to the `wwwroot` directory of the DotNet server application. All these changes are part of updating the application to use Vitelity for better performance and developer experience.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>