Update logo image and minor improvements in UI and backend
Updated the naming of the small bird logo image and its usage in "Footer.tsx" and "NavBar.tsx". This change reflects a more up-to-date version of the Used logo.

Modified the font styling and wordings in "Meal.tsx" to improve UI readability and user experience.

Cleaned up some syntax in "FetchUser.ts" and "Meal.tsx" to ensure code readability and further consistency across the project.

Changed from "FirstOrDefaultAsync" to "SingleAsync" method in "UserRepository.cs" in the backend, which will throw an error if there multiple users found. This is to ensure data integrity and only one user is returned for certain userid request.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>