Refactor Axios API calls and modify data handling
This commit includes both refactoring and changes in data handling structure. Moved Axios API calls and fetch functions from the context folder to a newly created FetchFunctions folder, to make the structure more intuitive and coherent. The content within FetchUser.ts, FetchUserContent.ts and index.ts were updated accordingly.

Updated the condition for setting 'isDataNotUndefined' state in MacroChartsLayout component. Used react-query in the MacroChartsLayout component to fetch meal data and handled loading and error states. Also, updated SingleMacroChart component to accept undefined data and modified the tickValues and tickFormat props on VictoryAxis.

Simplified async functions in FetchUser.ts and FetchUserContent.ts by replacing the use of await with then(). The async calls are now returning the axios instance directly. These changes are aiming to make promise-based async calls more readable and consistent.

Modified the NutrientService in dotnet backend to include a method for getting macros, updated its implementation and interface.

Other changes include updates in types and interfaces to align with the modifications in FetchUserContent.ts and EnergyAndMacroResponse.cs. The updates in DotNetServer csproj are small fixes on the directory structure of the included files.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>