Refactor nutrients calculation to use DTOs instead of entity
In the NorwegianFoodModule, the calculation logic that applies the quantity to the nutritional values of NorwegianFood has been moved from the entity to a new DTO, NorwegianFoodDto. This is to prevent any accidental DB updates caused by operating directly on the entity. Also, the calculation method GetTotalNutrientsResponse has been updated to accept a list of NorwegianFoodDto instead of NorwegianFood. Enhancements were made in the NorwegianFoodService and related mapping.

These changes align with good practice of not using the entity directly for operations that don't need to affect the database.
Also fixed possible null reference exceptions in EnergyAndMacroResponse and adjusted TotalNutrientsResponse to no longer set the CreatedAt property, which is unnecessary for its new role.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>