"Deleted unused components and cleaned up Hooks folder
The Food, Tables and SearchBar components were removed due to lack of use. Furthermore, redundant hooks in the hooks folder were deleted which has resulted in a simpler structure. Also, utilized the react-query library to fetch data in the Meal component, simplifying and reducing redundancy in the code. The general cleaning of the codebase will make future development easier and quicker."

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>