Refactor and relocate Calendar component to Charts
Relocated Calendar component to Charts directory as it fits better with the existing charting functionality in this directory. The Calendar.tsx was updated with new logic for calendar manipulation, creating a more interactive and intuitive user interface. Additionally, the Calendar component was integrated in TestPage.tsx, MacroChartsLayout.tsx, and SingleMacroChart.tsx. Furthermore, a Calendar.ts model was created to improve data management. Overall, these changes enhance the application's organization and user navigation experience.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>