Remove unused DotNetServer/wwwroot/assets/index-3a382489.js file
The client-side file "DotNetServer/wwwroot/assets/index-3a382489.js" has been removed because it is no longer used in the application. This include clean-up eliminates the need to load and maintain unnecessary JavaScript files, optimizing overall performance. Special attention has been paid to ensure that the removal of this file does not affect other parts of the application.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>