Update styling in LandingPage
Adjusted LandingPage styling for better responsiveness on different screen sizes. A responsive layout provides an optimal browsing experience for users across a wide range of devices. The div class has been modified to ensure the layout is full screen height. This improves the overall user experience and aesthetics of the page.

Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>